21.3 fb cds test fixes (#354)
* Test fix for missing value in x-axis and y-axis failure. Enabling the showHiddenVariable experimental flag.

* Setting the showHiddenVariable and resetting the variable at the end of the test.

* CDS Test fix

* Get the visible web elements on the page, don't use public locators.

* Moved the brushing tests into a separate test class (CDSVisualizationBrushingTest).
Removed unused variables from CDSVisualizationTest.
Replaced waitForElement(spinner) with a sleep.

* Opportunistic update.
Replacing assert with checker.
replacing a string check for blank.
