+++ test/src/org/labkey/test/pages/cds/CDSLoginPage.java
@@ -18,10 +18,17 @@ package org.labkey.test.pages.cds;
 import org.junit.Assert;
 import org.labkey.test.BaseWebDriverTest;
 import org.labkey.test.Locator;
+import org.labkey.test.WebDriverWrapper;
 import org.labkey.test.util.PasswordUtil;
+import org.openqa.selenium.WebElement;
+import java.util.List;
 
-import static org.labkey.test.pages.cds.CDSLoginPage.Locators.*;
-
+// Like the other CDS pages it would be nice to convert this page to a LabKeyPage object.
+// Even after that is done, one of the challenges with this and other CDS pages is that there might be several instances
+// of the same control present in the DOM but not visible. For example, depending upon the order the tests, this page
+// may have been hit before but in a slightly different way (valid user vs new user, etc...). The result is that the 'dialog'
+// that has the various login controls might exists multiple times on this page, but only one is visible. As a result
+// when the controls are referenced you need to make sure that it is visible control you have and not one of the hidden ones.
 public class CDSLoginPage
 {
     private final BaseWebDriverTest _test;
@@ -29,6 +36,8 @@ public class CDSLoginPage
     public CDSLoginPage(BaseWebDriverTest test)
     {
         _test = test;
+        // Wait until the signInButton is visible (not null) before returning.
+        WebDriverWrapper.waitFor(()->signInButton() != null, "Sign-In page did not load.", 5_000);
     }
 
     public void logIn()
@@ -38,20 +47,55 @@ public class CDSLoginPage
 
     public void logIn(String user, String password)
     {
-        Assert.assertTrue("Must agree to terms of use before logging in", termsCheckbox.findElement(_test.getDriver()).isSelected());
-        _test.setFormElement(emailField, user);
-        _test.setFormElement(passwordField, password);
-        _test.clickAndWait(signInButton);
+        Assert.assertTrue("Must agree to terms of use before logging in", termsCheckbox().isSelected());
+        _test.setFormElement(emailField(), user);
+        _test.setFormElement(passwordField(), password);
+        _test.clickAndWait(signInButton());
 
         _test.waitForElement(Locator.linkWithText("Logout"));
     }
 
-    public static class Locators
+    private WebElement findVisible(Locator locator)
     {
-        public static Locator emailField = Locator.id("email");
-        public static Locator passwordField = Locator.id("password");
-        public static Locator rememberMeCheckbox = Locator.checkboxById("remember-me-checkbox");
-        public static Locator termsCheckbox = Locator.checkboxById("tos-checkbox");
-        public static Locator signInButton = Locator.tagWithId("input", "signin");
+        // Yes this is ugly. The findElements will return only two or three controls, it is the check to see if it is
+        // visible that will take a while.
+        WebElement element = null;
+        List<WebElement> elements = locator.findElements(_test.getDriver());
+        for(WebElement el : elements)
+        {
+            if(el.isDisplayed())
+            {
+                element = el;
+                break;
+            }
+        }
+
+        return element;
     }
+
+    public WebElement emailField()
+    {
+        return findVisible(Locator.id("email"));
+    }
+
+    public WebElement passwordField()
+    {
+        return findVisible(Locator.id("password"));
+    }
+
+    public WebElement signInButton()
+    {
+        return findVisible(Locator.tagWithId("input", "signin"));
+    }
+
+    public WebElement rememberMeCheckbox()
+    {
+        return findVisible(Locator.checkboxById("remember-me-checkbox"));
+    }
+
+    public WebElement termsCheckbox()
+    {
+        return findVisible(Locator.checkboxById("tos-checkbox"));
+    }
+
 }
